% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadoi_fetch.r
\name{oadoi_fetch_}
\alias{oadoi_fetch_}
\title{Get open access status information.}
\usage{
oadoi_fetch_(doi = NULL, email = NULL)
}
\arguments{
\item{doi}{character vector,a DOI}

\item{email}{character verctor, tell oaDOI your email adress to get notified
if something breaks. It also helps oaDOI to keep track of usage!}
}
\value{
A tibble
}
\description{
In general, use oadoi_fetch instead. It calls this method, returning open
access status information from all your requests.
}
\examples{
\dontrun{
oadoi_fetch_(dois = c("10.1016/j.jbiotec.2010.07.030")
}
}

