% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadoiAddins.R
\name{roadoi_addin}
\alias{roadoi_addin}
\title{Find OA copies with RStudio addin}
\usage{
roadoi_addin()
}
\description{
An \href{https://rstudio.github.io/rstudioaddins/}{RStudio addin} to call
\code{oadoi_fetch()}. Shows up as "Find free full-texts" in the RStudio addin
menu.
}
\details{
The addin works as follows:

\enumerate{

\item Copy up to ten line-separated DOIs into the text area
\item Press the button "Run!"
\item Click on the links in the table to download full-text
}
}
