% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadoi_fetch.r
\name{oadoi_api_}
\alias{oadoi_api_}
\title{Post one request to access open access status information.}
\usage{
oadoi_api_(dois = NULL)
}
\arguments{
\item{dois}{character vector with dois}
}
\value{
A tibble
}
\description{
In general, use oadoi_fetch instead. It calls this method, returning open
access status information from all your requests. One request comprises up to
25 DOIs at a time.
}
\examples{
\dontrun{
oadoi_api_(dois = c("10.1016/j.jbiotec.2010.07.030",
"10.1186/1471-2164-11-245"))
}
}

