% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadoi_fetch.r
\name{oadoi_fetch}
\alias{oadoi_fetch}
\title{Fetch open access status information and full-text links from oaDOI}
\usage{
oadoi_fetch(dois = NULL, email = NULL, .progress = "none")
}
\arguments{
\item{dois}{character vector, search by a single DOI or many DOIs. The API is limited
to 10,000 requests per day. If you need more, get in touch with
team@impactstory.org.}

\item{email}{character verctor, tell oaDOI your email adress to get notified
if something breaks. It also helps oaDOI to keep track of usage!}

\item{.progress}{Shows the \code{plyr}-style progress bar. Options are "none", "text",
"tk", "win", and "time".  See \code{\link[plyr]{create_progress_bar}} for details
of each. By default, no progress bar is displayed.}
}
\value{
A tibble
}
\description{
This is the main function to retrieve comprehensive open acccess status
information from the oaDOI service. Please play nice with the API. At the
moment only 10k request are allowed per user and day.
For more info see \url{http://oadoi.org/api}.
}
\examples{
\dontrun{
oadoi_fetch("10.1016/j.shpsc.2013.03.020")
oadoi_fetch(dois = c("10.1016/j.jbiotec.2010.07.030",
"10.1186/1471-2164-11-245"))
}

}

