% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oadoi_fetch.r
\name{oadoi_fetch_}
\alias{oadoi_fetch_}
\title{Get open access status information.}
\usage{
oadoi_fetch_(doi = NULL, email = NULL)
}
\arguments{
\item{doi}{character vector,a DOI}

\item{email}{character vector, mandatory!
Unpaywall requires your email address,
  so that they can track usage and notify you when something breaks.
  Set email address in your `.Renviron` file with
  the option
  `roadoi_email` \code{options(roadoi_email = "najko.jahn@gmail.com")}.
  You can open your `.Renviron` file calling `file.edit("~/.Renviron")`.
  Save the file and restart your R session. To stop sharing your email
  when using rcrossref, delete it from your `.Renviron` file.}
}
\value{
A tibble
}
\description{
In general, use \code{\link{oadoi_fetch}} instead. It calls this
method, returning open access status information from all your requests.
}
\examples{
\dontrun{
oadoi_fetch_(doi = c("10.1016/j.jbiotec.2010.07.030"))
}
}
