% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadoi-package.r
\docType{package}
\name{roadoi-package}
\alias{roadoi-package}
\alias{roadoi}
\title{R Client for the oaDOI-API}
\description{
R Client for the oaDOI-API
}
\section{What is this client for?}{

roadoi interacts with the oaDOI data service, which links DOIs representing
scholarly works with open access versions.
}

\section{General usage}{

Use the \code{oadoi_fetch()} function in this package to get open access status
information and full-text links from oaDOI.
}

\section{Contribute}{

I would be very happy for people willing to contribute to this package.
It is important to keep in mind that oaDOI is still in early development,
which could affect the API and this client.
}

\keyword{package}
