% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadoi-package.r
\docType{package}
\name{roadoi-package}
\alias{roadoi-package}
\alias{roadoi}
\title{R Client for the Unpaywall-API}
\description{
What is this client for?:
roadoi interacts with the Unpaywall data service, which
links DOIs representing scholarly works with open access versions.
}
\details{
Use the \code{oadoi_fetch()} function in this package to get
open access status information and full-text links from Unpaywall.

You are welcome to contribute to this package. Use
GitHub's issue tracker for bug reporting and feature requests.

More details about the Unpaywall API:
\url{https://unpaywall.org/products/api}.
}
\keyword{package}
